Fix unresolved data with huge result set
on some case with many data / many services (mainly due to load activity on client site), $resource query is solved but data are not yet attached and mapped to resolved object. 
adding .$promise garanty that the service is done and data are fully mapped.

You can reproduce the case by displaying a search result on main page and doing a clean cache refresh.