Fix null pointer issue in hashCode method
Currently the generated hashCode method of the entities is not null safe. When a sql entity is created, and the hashCode method is called (for example, when adding it to a HashSet), the method throws a null pointer exception since the id field is not yet set.
To fix the issue I have changed the hashCode method to use the java Objects.hashCode helper method that by default is null safe. It is also slightly more readable.
I have also changed the equals method to use the Objects helper method to make it slightly more readable.