README.md with build instructions
create README.md with build instructions for Grunt and Gradle

This is a first pass combining things from the JHipster Mini-Book and
the current Angular/Grunt project I’m working on for a client. In our
environment, we have a remote backend on GAE, so it’s easy to
build/test the client w/o having a backend running. This section could
probably be improved to show how to test with ‘gradlew bootRepackage’.

This is just a first pass and meant for improvement to handle the
various options.

#1880