Fixed #2552
"Create new <Entity>"button is left aligned (padding removed)
Search bar is now an input group (input followed by search icon) and left-aligned.
Search bar text "Search a <Entity>" is now in a tooltip.
The text "Search a <Entity>" was not localized, added english.
When shrinking the window (and search is enabled), the text "Create new
button" is removed  on smaller screen sizes, text is available as a tooltip.
The "View", "Edit", and "Delete" buttons are now inside button group.
Table data text is vertifcally-aligned in the middle now.
Updated language files.
Restored ng-click that was removed by misstake.

Fix #2552