Define a constant instead of duplicating this literal "user" 4 times.
Define a constant instead of duplicating this literal "baseUrl" 3 times.
Define a constant instead of duplicating this literal "user" 4 times.

SONAR: Duplicated string literals make the process of refactoring error-prone, since you must be sure to update all occurrences. On the other hand, constants can be referenced from many places, but only need to be updated in a single place.