Add support for html tags in UI alerts
* Add support for html tags in UI alerts

Added ng-bind-html to alert.directive.js to support formatting bootstrap alert messages with HTML tags (bold, italics, etc.).

* Fixed new password validation message

The app checks if a new password is at least 4 characters long, but the message says 5. Fixed for EN.

* Fixed new password validation message (CS)

The app checks if a new password is at least 4 characters long, but the message says 5. Fixed for CS.