involves AuthorizedFeignClient and makes the usage of feign clients possible in general
In the previous version of this PR, every feign client in the project would be intercepted
with the OAuth2RequestInterceptor. Removing this behaviour makes it possible, to configure
any kind of feign client. To fill the lack of a preconfigured feign client, able to cleanely
communicate with JHipsters UAA, the developer may use `@AuthorizedFeignClient` instead.

Note: due to a annotation bug, only `@AuthorizedFeignClient(name = "otherapp")` works, but not
`@AuthorizedFeignClient("otherapp")`

PR #3662