Improve error message for parsing JDL's
When there is an exception thrown that is of type `Error` the error handling prints "[object Object]". This change checks if there is a message property and prints that, otherwise it assumes its a string. This is the one place I was having issue with, but the same problem exists anywhere that pattern is used.