Change Eureka Environment variable to valid character
The “EUREKA_CLIENT_SERVICE-URL_DEFAULTZONE” environment variable specified in the _app.yml file contains an invalid hyphen(-) character. When the container is started, this variable is not set. This change replaces the hyphen with an underscore, which Spring Boot can still read due to its relaxed binding rules. This change aligns the _app.xml file it to the Kubernetes+Openshift versions of this template.