Use the real fs when reading entity JSON files
Normally `this.fs` is yeoman's in-memory filesystem (mem-fs) so once
file is read from disk it's not re-read.

Using jhipster-core function `readEntityJSON` means we always read from
disk and have most up to date and consistent state.

This came up when implementing a change in `import-jdl` generator where
entities needed to be re-read after they've been changed by the import.