Ignore JHipster package.json scripts during upgrade
Npm scrpits of generator-jhipster are useful for global installation but they are not required in teh context of an upgrade where the focus is on code and depenencies.

These scripts can make the upgrade fail for bad reasons llike issue #6162 about tabtab. 

This change add `--ignore-scripts`option  to yarn or npm install during upgrade.