fix tests
Add info.project.version property in test configuration for apps not using serviceDiscovery.
This is needed as the starting the tests context for those apps will load LoggingConfiguration
which require this property to be set. This is not required for apps using eureka or consul
service discovery as the LoggingConfiguration is only loaded for those apps when service discovery
is active.