6668 -  fix the logstash-appender stopping working issues (#6679)
* fix the logstash-appender stopping working issues 

In the SpringCloudConfig, Spring will refresh its configuration even the app is already running.  Spring will call the `LoggerContext` to refresh all the logger config. which will `StopAndReset` the log config. it will stop all the logappender and remove all the `LoggerContextListener`.   After that, `LoggerContext `will load the new configuration based on the configuration file (logback-spring.xml) . After that, Spring will call the configuration class ...... , but for this case, the bean is singleton and already been created. 
In this case, the above `LoggingConfiguration`  won't be created again (even the `logstash-appender` has been stopped).

We add the `@RefreshScope` here to force Spring to recreate this bean when there is configuration refresh

* change the condition for enable 'RefreshScope'

* import the RefreshScope when the `serviceDiscovery` is enabled