Allow Spring boot Scheduled tasks to have multiple threads
The Async configuration currently only allows us to set the async pool size and not the scheduler pool size.
This results in spring scheduled tasks having only one thread to execute all it's tasks.

The proposed change allows us to have the number of scheduler tasks to be the size of the core pool.

Ideally we should have one more configuration value in application.yml which will dictate the size of the executor pool.

See: https://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/scheduling/annotation/EnableScheduling.html

"By default, will be searching for an associated scheduler definition: either a unique TaskScheduler bean in the context, or a TaskScheduler bean named "taskScheduler" otherwise; the same lookup will also be performed for a ScheduledExecutorService bean. If neither of the two is resolvable, a local single-threaded default scheduler will be created and used within the registrar. "