When junit microservices configured with uaa, I get error in the stack trace.  The test pass but this stacktrace doesn't look good.
With junit, when we mock a remote server, UaaSignatureVerifierClient constructor receive a null value for the restTemplate.
So we first need to accept null in that case.

Then we need to change the order of the @SpringBootTest classes.
We need to set the SecurityBeanOverrideConfiguration.class as the second in the list so it's initialized first.