pass app version to NPM as env var instead of parsing it from build file
Currently, the webpack `DefinePlugin` sets the `VERSION` by parsing it from either `build.gradle`
(in case of a Gradle build) or from `pom.xml` (in case of a Maven build). The problem with this is
that if the version is not statically set, but rather dynamically calcualted (e.g. using
https://github.com/palantir/gradle-git-version to determine it from the Git repo), the parsing
breaks.

Thus, this PR changes this by passing the version to show in the frontend from the server build
tool (Gradle/Maven) as an environment variable to the NPM webpack task. This is more robust and
does not rely on parsing the Gradle/Maven file. In the Maven case, it also removes the dependency
on xml2js. The downside is that if only `npm start` is run, instead of the version, `UNKNOWN` is
shown.

I think the advantages (allowing dynamic version in Gradle, removing parsing code) outweigh this
downside.

I tested this by
1. generating Maven and Gradle projects
2. generating Angular and React clients
3. running it using `dev` vs. `prod` profiles
4. running the server build tool vs. `npm start`

(but not all combinations of the above :-).