allow using Gradle project version in jib configuration (#9985)
We decided to not use `latest` as the tag for the Docker image built by jib, but the Gradle project version. While I think changing the default Docker image tag is not a good change for JHipster in general, moving the project version up allows it to be referenced by Jib. Not knowing too much about Gradle project configuration before, we stumbled over this and I want to save others this learning.