use default link colors of Bootswatch theme
With c55634ed99317aedebf5efad441a3a3e87f3f23b, it was decided to override the link color of Bootstrap/the Bootswatch theme for passing the Lighthouse accessibility audit.

I think this is a bad idea since 1. the picked color might not match the color scheme of the Bootswatch theme and 2. it should be the responsibility of the theme to consider accessibility concerns and not JHipster building on top of this.

This commit only applies the link color when no Bootswatch theme is chosen. In this case, the Bootstrap default theme is used and Bootstrap recommends to 'manually modify/extend these default colors [e.g. the link color] to ensure adequate color contrast ratios': https://getbootstrap.com/docs/4.1/getting-started/accessibility/#color-contrast.