make variable overrides in _bootstrap_variables.scss change components (#10119)
Currently, the `_bootstrap_variables.scss` file is not included into `vendor.scss`. Thus, the overriden variables in there are not affecting the bootstrap components. This commit fixes that. Now, the variables defined in this file influence the bootstrap component colors and other styles.

Since `_bootstrap-variables.scss` is also present (but being empty) when generating without a Bootswatch client theme, I think it still makes sense to also include it in this case. But in order to allow overriding variables already set in the `variables.scss` file of the Bootswatch theme, `_bootstrap-variables.scss` needs to be imported after this file.