Upgrade Travis pipeline to OpenJDK11
I've removed the jdk option of travis; this I believe is useless since it's only used when used in conjunction with language: java. 

Also I've changed the linux distro to bionic since the default distro used by travis seems to be trusty and that doesn't have openjdk11 preinstalled (bionic does). This I believe makes the travis pipeline cleaner so that we don't have to manually install stuff on it and also could get rid of the jdk_switcher use oraclejdk8 part (which seems redundant anyways).