Improve shared libs module in Angular
If currently using CookieService from ngx-cookie in some lazy loaded module which imports SharedModule (developer modifies code and this modification uses CookieService) then new instance of the CookieService is created. But the idea of the ngx-cookie library is that CookieService should be singleton. This change makes CookieService to singleton.
This also removes entirely imports part from shared libs module because shared libs module is not declaring any component, directive or pipe and therefore no imports needed.