Better exception/log for MailService
Only `MessagingException` should be thrown by the code, so we should only catch them instead of `Exception`

Testing `isDebugEnabled` while the message is logged at warn level is very strange. If level is configured at DEBUG, no message will be logged because of the `warn` level of the logs. So the `else` branch is useless, and only the `if` branch is useful. Also, logging only the message of the exception is bad practice.