introduce dynamic storage provisioning in k8s sub-gen
with this feature, all stateful services (such as `Deployment`, `StatefulSet` for databases) will use persistent volume claims instead of `emptyDir`. There are two new options for the user, so it is possible to enable dynamic storage provisioning and specify a custom storage class.

Another update in context with this PR, is that `StatefulSet`s are no longer generated with `volumeClaimTemplate` by default, as it was before. Instead, the data dirs are either `emptyDir`, if dynamic storage provisioning is disabled, or we use volume claim templates with the same configuration as for the deployments.

close #10804