Change cli:utils.done from fatal to error:
Behavior changes:
import-jdl.js: when importing jdl entities, the process will not quit on
jdl errors.

run-yeoman-process.js: child process will not be forced to quit on
error.

cli.js: commands will not be forced to quit on error.
This seems to be cli last step, so no behavior changes.