Breaking change: Introduced the JDLApplication and modified the JDL app factory
Replaced the previous JDL application classes by the single JDLApplication.
The goal is to make changes easier to do, and reduce complexity as inheritance
made changes delicate and in more than one file.
Now, for each application type, a function exists to get the correct config as
well as default configuration.
Refactored the impacted files to make it work again.
The correct way to create an application now is to:
  - either provide a full configuration to the JDLApplication
  - or use the JDL app factory and provide custom options

Removed:
  - AbstractJDLApplication and its 4 concrete classes
    - Poor design choices as inheritance wasn't the best choice (see previous commit for more info)