Added support for decimal values in the JDL
For validations:
  - min & max
  - also for constants

Even if the decimal values are allowed at parsing-time for maxlength, etc.,
the validation validator will take care of throwing an exception (with a nice
message) if such a value is used.

Also reverts the code of the previous commit, because stringifying the
decimal or integer value of a validation is also done at parsing-time.

Fixes #423