fix WebFlux integration tests
With Spring test support applying all our real world filters, this means that for reactive applications also `SpaWebFilter` will now be applied.
This in turn will redirect the request to `/index.html` for `ExceptionTranslatorIT` which will then fail the test. To fix this, we move `ExceptionTranslatorTestController` to a new path, namely under `/api`. Unfortunately this means that we need to add CSRF support for certain scenarios (i.e. authentication type `session`), since CSRF protection is added for `/api`.
Furthermore I unified CSRF conditions in tests by introducing a new property within `index.js` instead of adding the complex if-statement to the tests.