remove endpoint `/api/authenticate` for `oauth2` config
As explained by @mraible the endpoint should be useless for `oauth2` configurations. Due to a customization in `SecurityConfiguration` the endpoint could so far not be queried by an unauthenticated user, while it could be queried by an unauthenticated user for all other configurations. To remove complexity we have decided to remove this endpoint altogether for `oauth2`, thus making it unnecessary to fix/adjust the test, but simply remove it. This will also make the code easier to understand in the future.