set `stateless` to `false` within OAuth2 config for `uaa` tests
By default `stateless` is set to `true` which causes the `SecurityContext` set up by the test to be cleared within the filter chain. This in turn destroys the authentication set up by the test. To resolve this problem we add a specific test configuration that configures the OAuth2 filter to `stateless = false` so the `SecurityContext` is preserved for the test and the test in turn behaves consistently to the other configurations.