Simplified and hopefully fixed e2e test
Why simplified: because the deletion is actually done in steps. First, the
user clicks the entry deletion button, then the modal appears which hides the
table, the user clicks the button and the request is sent to the backend.
After that, the table is displayed and the toaster can appear before the
table is updated.

In the test, we ignore this fact and wait for the modal to disappear but
not for the table to be updated.
The test basically now says: okay we click the button, wait for the modal to
go away and then we wait for the table to update because there won't be any
entry after that.
If there's a backend issue, the wait will be over and the test will fail
because the last entry won't be deleted.