fix: mistakes in logic concerning dependabot CI exclusions
This changes the logic in dependabot pull request CI skipping. Current logic was created with the hope dependabot creates a label on the PR creation. But unfortunately dependabot adds the label after the PR is created by which time the CI has already been triggered.

Related to https://github.com/jhipster/generator-jhipster/issues/11962