Prevented commit additions when no JDL dist file have changed
Used the simple-git package to programmatically use git inside a script and
created a script that:
  - checked the dist file diffs
  - add the files
  - commit them
If no file has changed, then no commit is done for a NPM prepare step.
The prepare steps does:
```
Run both BEFORE the package is packed and published,
and on local npm install without any arguments.
This is run AFTER prepublish, but BEFORE prepublishOnly
```
Only does a commit if a JDL dist file has changed.