[angular] Reduce event manager importance and add stricter types
As there is not possible to apply strict type checking if using event manager then this usage should not be encouraged. So moved this to util subfolder instead of separate folder and put model to the same file where is service to reduce event manager visibility and importance.
Added also strict types instead of any to event manager. Now consumers must do type cast - so if they use event manager they take a conscious risk.
Added block "unreleased files and folders cleanup for v7 developers" to cleanup to delete files if developers regenerate with newer version in v7 development cycle.