fix race condition during react protractor e2e tests
The e2e test changes the admin password, then attempts to change
the password back to 'admin'.  Before the request was completed,
the test would finish and the window reloads in preparation for the next test.

This fix waits for the password to be changed, then moves on to the next test.

related to https://github.com/jhipster/generator-jhipster/issues/13361#issuecomment-752270050