fix export/import of otherModules property
Exporting Application JDL containing `otherModules` (for example when a blueprint is used) results in a JDL with `otherModules [[Object object]]` which fails when attempting to import.

This PR aligns the logic of the `otherModules` property with the `blueprints` property, which has the same format. With this change, apps generated with blueprints can export valid JDL.