improve check for existing heroku app (#13718)
If the `herokuAppName` is set, we check for the existence of the app.  This only works if it has previously been deployed from the same device and directory (it checks for the git remote).  If a developer clones the project from a repository then tries to run `jhipster heroku`, it will not find the existing app and reset the `herokuAppName` in the `yo-rc.json`

By passing the `herokuAppName` as a param to `heroku apps:info`, we can find out if the app exists. We also find out if the user has access to the application, otherwise an error is returned.