fix test when nested jpaDerivedIdentifier is used
when a relationship is used with the jpaDerivedIdentifier option and the
related entity also has a relationship with that option set the
`updateEntityMapsIdAssociationWithNewId` test fails. This
is caused by the `createUpdatedEntity` reusing the needed parent
entity. This can't work because a parent entity, which already has a child,
can't be used as the parent for a new entity. The reason being
jpaDerivedIdentifier enforces that parent and child share the same id,
which can't be fulfilled with a reused parent entity.
Remove the reuse of the parent entity in `createUpdatedEntity` when the
entity uses jpaDerivedIdentifier to fix the test.

Fixes #17906