fix @vue/eslint-config-prettier dependency
Currently, the newest version of `@vue/eslint-config-prettier`
does not support prettier 3 because it depends on an older
version of `eslint-plugin-prettier` that does not
support prettier 3. This override forces `@vue/eslint-config-prettier`
to use the newer `eslint-plugin-prettier`.
When workspaces are used the override must be defined in the root package.json
because overrides in the workspaces package.json are not supported
by npm.
This workaround can be removed when `@vue/eslint-config-prettier` supports
prettier 3.