fix(generator): default the maven-resources-plugin definition to the spring-boot-parent definition
Update the generator templates for Gradle and Maven to use the Spring default delimiter of the
commercial-at sign (`@`). Remove the JHipster configuration of the delimiter and default to Spring
Boot Parent's configuration and allow it to inherit. This will properly scope the config/bootstrap*.yml and
allow Spring Boot's root resources definition to handle the **/application*.yml files.

Fixes #25979